requirements.txt
Flask==3.0.0
pymongo==4.6.0
python-dotenv==1.0.0
flask-cors==4.0.0
email-validator==2.1.0
bcrypt==4.1.2
jsonschema==4.20.0
